feat(Graph, linter config): Updated graph component to render data properly
Updates to the heroLineup are stored in the data array which is passed to the rechart components.
Each time a new hero is added, its stats are included on top of the previous states, so that rather
than having a bunch of overlapping polar graph shapes, the graph expands outward.

none